pillar.show_top needs to be a runner, not a module function
